+++ plugin_ide.ui/src/com/googlecode/goclipse/ui/editor/actions/RunGoFixOperation.java
@@ -15,6 +15,7 @@ import melnorme.lang.ide.ui.actions.AbstractEditorHandler;
 import melnorme.utilbox.core.CommonException;
 
 import org.eclipse.core.runtime.CoreException;
+import org.eclipse.ui.IWorkbenchPage;
 import org.eclipse.ui.texteditor.ITextEditor;
 
 import com.googlecode.goclipse.core.GoCore;
@@ -22,12 +23,14 @@ import com.googlecode.goclipse.tooling.env.GoEnvironment;
 
 public class RunGoFixOperation extends AbstractEditorGoToolOperation {
 	
-	public static final AbstractEditorHandler handler = new AbstractEditorHandler() {
-		@Override
-		public void runOperation(ITextEditor editor) {
-			new RunGoFixOperation(editor).executeAndHandle();
-		}
-	};
+	public static AbstractEditorHandler getHandler(IWorkbenchPage page) {
+		return new AbstractEditorHandler(page) {
+			@Override
+			public void runOperation(ITextEditor editor) {
+				new RunGoFixOperation(editor).executeAndHandle();
+			}
+		};
+	}
 	
 	public RunGoFixOperation(ITextEditor editor) {
 		super("Run 'go fix'", editor);