+++ plugin_ide.ui/src/com/googlecode/goclipse/ui/editor/actions/RunGoFmtOperation.java
@@ -14,6 +14,7 @@ import static melnorme.utilbox.core.CoreUtil.listFrom;
 import melnorme.lang.ide.ui.actions.AbstractEditorHandler;
 
 import org.eclipse.core.runtime.CoreException;
+import org.eclipse.ui.IWorkbenchPage;
 import org.eclipse.ui.texteditor.ITextEditor;
 
 import com.googlecode.goclipse.core.GoEnvironmentPrefs;
@@ -21,12 +22,14 @@ import com.googlecode.goclipse.tooling.env.GoEnvironment;
 
 public class RunGoFmtOperation extends AbstractEditorGoToolOperation {
 	
-	public static final AbstractEditorHandler handler = new AbstractEditorHandler() {
-		@Override
-		public void runOperation(ITextEditor editor) {
-			new RunGoFmtOperation(editor).executeAndHandle();
-		}
-	};
+	public static AbstractEditorHandler getHandler(IWorkbenchPage page) {
+		return new AbstractEditorHandler(page) {
+			@Override
+			public void runOperation(ITextEditor editor) {
+				new RunGoFmtOperation(editor).executeAndHandle();
+			}
+		};
+	}
 	
 	public RunGoFmtOperation(ITextEditor editor) {
 		super("Run 'gofmt'", editor);